Fixed decoding double bug found by random tests

+++ src/org/rascalmpl/library/lang/json/io/JsonValueReader.java
@@ -123,7 +123,7 @@ public class JsonValueReader {
         try {
           switch (in.peek()) {
             case NUMBER:
-              return vf.real(in.nextInt());
+              return vf.real(in.nextDouble());
             case STRING:
               return vf.real(in.nextString());
             case NULL:
